Update experiment trial workflow
Update experiment trial workflow to what should be about the final
version we need.  This stroop replicatione experiment does the following

- Prompts and records participant number, session number and posture condition
- Runs 1 block of 36 practice trials
- Gives feedback in form of sound and reaction time result for practice
- Has a limit of 1.5 seconds to respond on trials
- Runs 2 blocks of 36 trials for real data
- Give sound feedback but no other information on the real trials.

It is assumed that the experimenter will run subject in 1 posture first.
Then reposition subject for new posture and run a second time with the
same participant id.  Thus we will end up with 2 separate files for
each participant, containing the sitting and standing posture data
respectively.